fixed database association, everything seems to be talking. Also, assigned delete features to admin and created an admin login in postgres

+++ controllers/coffeecontroller.js
@@ -3,35 +3,68 @@ const router = express.Router();
 const sequelize = require('../db');
 const validateSession = require('../middleware/validate-session');
 // var Coffee = sequelize.import('../models/coffee');
+var Comment = sequelize.model('comment');
 var Coffee = sequelize.model('coffee');
 
 router.post('/create', validateSession, (req, res) => {
+  let fullName = req.user.firstName + ' ' + req.user.lastName;
   if (!req.errors) {
       const coffeeRequest = {
         name: req.body.name,
         location: req.body.location,
         favoriteDrink: req.body.favoriteDrink,
         note: req.body.note,
-        owner: req.user.id
+        userId: req.user.id,
+        owner: fullName
       }
       Coffee.create(coffeeRequest)
-      .then(coffee => res.status(200).json(coffee))
-      .catch(err => res.json(req.errors))
+      .then(newCoffee => {
+        res.json({
+          coffee: newCoffee
+        })
+      })
+      .catch(err => {
+        res.json({
+          error: err.message
+        })
+      })
   } else {
-    res.status(500).json(req.errors)
+    console.log('It didnt even make it');
   }
 })
 
 router.get('/:id', validateSession, (req, res) => {
-  Coffee.findOne({ where: { id: req.params.id },
-    Include: [
+  Coffee.findOne({ 
+    where: { id: req.params.id },
+    include: [
       {
       model: Comment,
       where:{ coffeeId: req.params.id }
       }
     ]
   })
-    .then(coffee => res.status(200).json(coffee))
+    .then(coffee => {
+      // If no comments
+      if (coffee === null) {
+        Coffee.findOne({
+          where: {id: req.params.id}
+        })
+        .then(coffee => {
+          res.json({
+            coffee: coffee
+          })
+        })
+        .catch(err => {
+          res.json({
+            error: err.message
+          })
+        })
+      } else {
+        res.json({
+          coffee: coffee
+        })
+      }
+    })
     .catch(err => res.status(500).json({ error: err}))
 })
 
@@ -54,11 +87,27 @@ router.put('/:id', validateSession, (req, res) => {
 
 router.delete('/:id', validateSession, (req, res) => {
     if (!req.errors) {
-      Coffee.destroy({ where: { id: req.params.id }})
+      Coffee.destroy({ where: { userId: req.user.id }})
       .then(coffee => res.status(200).json(coffee))
       .catch(err => res.json(req.errors))
   } else {
     res.status(500).json(req.errors)
+  }
+})
+
+router.delete('/:id/admin', validateSession, (req, res) => {
+  if (req.user.role === 'admin') {
+    if (!req.errors) {
+      Coffee.destroy({ where: { id: req.params.id }})
+      .then(coffee => res.status(200).json(coffee))
+      .catch(err => res.json(req.errors))
+    } else {
+      res.status(500).json(req.errors)
+    }
+  } else {
+    res.json({
+      message: 'You do not have permission to access this data!'
+    })
   }
 })
