+++ controllers/commentcontroller.js
@@ -5,13 +5,15 @@ const validateSession = require('../middleware/validate-session');
 // var Comment = sequelize.import('../models/comment');
 var Comment = sequelize.model('comment');
 
-router.post('/create', validateSession, (req, res) => {
+router.post('/coffee/:coffeeId/comment/create', validateSession, (req, res) => {
+  let fullName = req.user.firstName + ' ' + req.user.lastName
   if (!req.errors) {
       const commentRequest = {
-        name: req.body.name,
+        name: fullName,
         comment: req.body.comment,
         rating: req.body.rating,
-        owner: req.user.id
+        owner: req.user.id,
+        coffeeId: req.params.coffeeId
       }
       Comment.create(commentRequest)
       .then(comment => res.status(200).json(comment))
@@ -27,13 +29,13 @@ router.post('/create', validateSession, (req, res) => {
 //     .catch(err => res.status(500).json({ error: err}))
 // })
 
-router.get('/owner', validateSession, (req, res) => {
-    Comment.findAll({ where: { owner: req.user.id }})
-        .then(comment => res.status(200).json(comment))
-        .catch(err => res.status(500).json({error: err}))
-})
+// router.get('/comment/:id', validateSession, (req, res) => {
+//     Comment.findAll({ where: {  }})
+//         .then(comment => res.status(200).json(comment))
+//         .catch(err => res.status(500).json({error: err}))
+// })
 
-router.put('/:id', validateSession, (req, res) => {
+router.put('/comment/:id', validateSession, (req, res) => {
   if (!req.errors) {
     Comment.update(req.body.comment, { where: { id: req.params.id }})
       .then(comment => res.status(200).json(comment))
@@ -43,13 +45,29 @@ router.put('/:id', validateSession, (req, res) => {
   }
 })
 
-router.delete('/:id', validateSession, (req, res) => {
+router.delete('/comment/:id', validateSession, (req, res) => {
     if (!req.errors) {
-      Comment.destroy({ where: { id: req.params.id }})
+      Comment.destroy({ where: { userId: req.user.id }})
       .then(comment => res.status(200).json(comment))
       .catch(err => res.json(req.errors))
   } else {
     res.status(500).json(req.errors)
+  }
+})
+
+router.delete('/comment/:id/admin', validateSession, (req, res) => {
+  if (req.user.role === 'admin') {
+    if (!req.errors) {
+      Comment.destroy({ where: { id: req.params.id }})
+      .then(comment => res.status(200).json(comment))
+      .catch(err => res.json(req.errors))
+    } else {
+      res.status(500).json(req.errors)
+    }
+  } else {
+    res.json({
+      message: 'You do not have permission to access this data!'
+    })
   }
 })
